+++ sc/social/like/interfaces/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa
-
-from socialikes import IHelperView
-from socialikes import ISocialLikeLayer
-from socialikes import ISocialLikes