+++ sc/social/like/logger.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from sc.social.like.config import PROJECTNAME
+
+import logging
+
+logger = logging.getLogger(PROJECTNAME)