+++ sc/social/like/plugins/twitter/browser.py
@@ -1,9 +1,11 @@
 # -*- coding:utf-8 -*-
-from Products.CMFCore.utils import getToolByName
+from plone import api
+from plone.api.exc import InvalidParameterError
 from Products.CMFPlone.utils import safe_unicode
 from Products.Five import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.PythonScripts.standard import url_quote
+from sc.social.like.interfaces import ISocialLikeSettings
 from sc.social.like.utils import get_language
 from urllib import urlencode
 from zope.component import getMultiAdapter
@@ -11,10 +13,7 @@ from zope.component import getMultiAdapter
 
 class PluginView(BrowserView):
 
-    enabled_portal_types = []
-    typebutton = ''
-    twitter_enabled = ''
-    twittvia = ''
+    twitter_enabled = False
     language = 'en'
 
     metadata = ViewPageTemplateFile('templates/metadata.pt')
@@ -22,11 +21,11 @@ class PluginView(BrowserView):
     link = ViewPageTemplateFile('templates/link.pt')
 
     def __init__(self, context, request):
-        super(PluginView, self).__init__(context, request)
-        pp = getToolByName(context, 'portal_properties')
-
         self.context = context
         self.request = request
+        # FIXME: the following could rise unexpected exceptions
+        #        move it to a new setup() method
+        #        see: http://docs.plone.org/develop/plone/views/browserviews.html#creating-a-view
         self.portal_state = getMultiAdapter((self.context, self.request),
                                             name=u'plone_portal_state')
         self.portal = self.portal_state.portal()
@@ -34,26 +33,38 @@ class PluginView(BrowserView):
         self.portal_title = self.portal_state.portal_title()
         self.url = context.absolute_url()
         self.language = get_language(context)
-        self.sheet = getattr(pp, 'sc_social_likes_properties', None)
-        if self.sheet:
-            self.typebutton = self.sheet.getProperty('typebutton', '')
-            self.twittvia = self.sheet.getProperty('twittvia', '')
         self.urlnoscript = (
             u'http://twitter.com/home?status=' +
             url_quote(u'{0} - {1} via {2}'.format(
                 safe_unicode(self.context.title),
                 self.context.absolute_url(),
-                self.twittvia)
+                self.via)
             )
         )
+
+    @property
+    def typebutton(self):
+        record = ISocialLikeSettings.__identifier__ + '.typebutton'
+        try:
+            return api.portal.get_registry_record(record)
+        except InvalidParameterError:
+            return ''
+
+    @property
+    def via(self):
+        record = ISocialLikeSettings.__identifier__ + '.twitter_username'
+        try:
+            return api.portal.get_registry_record(record)
+        except InvalidParameterError:
+            return ''
 
     def share_link(self):
         params = dict(
             text=self.context.Title(),
             url=self.context.absolute_url(),
         )
-        if self.twittvia:
-            params['via'] = self.twittvia
+        if self.via:
+            params['via'] = self.via
 
         url = 'https://twitter.com/intent/tweet?' + urlencode(params)
         return url