+++ sc/social/like/tests/test_plugin_linkedin.py
@@ -1,12 +1,14 @@
 # -*- coding: utf-8 -*-
 from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
-from sc.social.like.controlpanel.likes import LikeControlPanelAdapter
+from plone.registry.interfaces import IRegistry
+from sc.social.like.interfaces import ISocialLikeSettings
 from sc.social.like.interfaces import ISocialLikeLayer
-from sc.social.like.plugins.linkedin import browser
 from sc.social.like.plugins.interfaces import IPlugin
+from sc.social.like.plugins.linkedin import browser
 from sc.social.like.testing import INTEGRATION_TESTING
 from zope.component import getUtilitiesFor
+from zope.component import getUtility
 from zope.interface import alsoProvides
 
 import unittest
@@ -54,9 +56,12 @@ class PluginViewsTest(unittest.TestCase):
 
     def setUp(self):
         self.portal = self.layer['portal']
-        self.adapter = LikeControlPanelAdapter(self.portal)
         setRoles(self.portal, TEST_USER_ID, ['Manager'])
         self.setup_content(self.portal)
+
+        self.registry = getUtility(IRegistry)
+        self.settings = self.registry.forInterface(ISocialLikeSettings)
+
         alsoProvides(self.portal.REQUEST, ISocialLikeLayer)
         self.plugins = dict(getUtilitiesFor(IPlugin))
         self.plugin = self.plugins[name]
@@ -83,8 +88,8 @@ class PluginViewsTest(unittest.TestCase):
     def test_privacy_plugin_view_html(self):
         plugin = self.plugin
         portal = self.portal
-        properties = portal.portal_properties.sc_social_likes_properties
-        properties.do_not_track = True
+        self.settings.do_not_track = True
+
         plugin_view = plugin.view()
         view = portal.restrictedTraverse(plugin_view)
         html = view.link()
@@ -107,7 +112,6 @@ class PluginViewsTest(unittest.TestCase):
         self.assertEqual(view.typebutton, 'right')
 
         # Change to vertical
-        adapter = self.adapter
-        adapter.typebutton = 'vertical'
+        self.settings.typebutton = 'vertical'
         view = portal.restrictedTraverse(plugin_view)
         self.assertEqual(view.typebutton, 'top')